+++ tools/emterpretify.py
@@ -35,7 +35,6 @@ ZERO = False
 ASYNC = False
 ASSERTIONS = False
 PROFILING = False
-SWAPPABLE = False
 FROUND = False
 ADVISE = False
 MEMORY_SAFE = False
@@ -799,9 +798,6 @@ if __name__ == '__main__':
   if len(sys.argv) >= 6:
     SYNC_FUNCS.update(read_json_list(sys.argv[5]))
 
-  if len(sys.argv) >= 7:
-    SWAPPABLE = int(sys.argv[6])
-
   if ADVISE:
     # Advise the user on which functions should likely be emterpreted
     data = shared.Building.calculate_reachable_functions(infile, list(SYNC_FUNCS))
@@ -831,7 +827,7 @@ if __name__ == '__main__':
 
   BLACKLIST = set(list(BLACKLIST) + extra_blacklist)
 
-  if DEBUG or SWAPPABLE:
+  if DEBUG:
     orig = infile + '.orig.js'
     shared.logger.debug('saving original (non-emterpreted) code to ' + orig)
     shutil.copyfile(infile, orig)